Refactored AutoFlushingStream
