+++ src/Razor/src/Microsoft.VisualStudio.LanguageServerClient.Razor/RazorLanguageServerClient.cs
@@ -96,51 +96,5 @@ namespace Microsoft.VisualStudio.LanguageServerClient.Razor
         }
 
         public Task AttachForCustomMessageAsync(JsonRpc rpc) => Task.CompletedTask;
-
-        private class AutoFlushingStream : Stream
-        {
-            private readonly Stream _inner;
-
-            public AutoFlushingStream(Stream inner)
-            {
-                _inner = inner;
-            }
-
-            public override bool CanRead => _inner.CanRead;
-
-            public override bool CanSeek => _inner.CanSeek;
-
-            public override bool CanWrite => _inner.CanWrite;
-
-            public override long Length => _inner.Length;
-
-            public override long Position { get => _inner.Position; set => _inner.Position = value; }
-
-            public override long Seek(long offset, SeekOrigin origin) => _inner.Seek(offset, origin);
-
-            public override void SetLength(long value) => _inner.SetLength(value);
-
-            public override void Flush() => FlushAsync(CancellationToken.None).GetAwaiter().GetResult();
-
-            public override int Read(byte[] buffer, int offset, int count) => ReadAsync(buffer, offset, count, CancellationToken.None).GetAwaiter().GetResult();
-
-            public override void Write(byte[] buffer, int offset, int count) => WriteAsync(buffer, offset, count, CancellationToken.None).GetAwaiter().GetResult();
-
-            public override async Task FlushAsync(CancellationToken cancellationToken)
-            {
-                await _inner.FlushAsync(cancellationToken).ConfigureAwait(false);
-            }
-
-            public override async Task<int> ReadAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-            {
-                return await _inner.ReadAsync(buffer, offset, count).ConfigureAwait(false);
-            }
-
-            public override async Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
-            {
-                await _inner.WriteAsync(buffer, offset, count).ConfigureAwait(false);
-                await FlushAsync().ConfigureAwait(false);
-            }
-        }
     }
 }
